Remove AbstractCRUD and CRUDElement and move delete posibility into GenericListElement.
