--- spec/AdminPanel/Symfony/AdminBundle/Admin/CRUD/AbstractCRUDSpec.php
@@ -1,96 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace spec\AdminPanel\Symfony\AdminBundle\Admin\CRUD;
-
-use AdminPanel\Symfony\AdminBundle\Exception\RuntimeException;
-use PhpSpec\ObjectBehavior;
-use Prophecy\Argument;
-
-class AbstractCRUDSpec extends ObjectBehavior
-{
-    public function let()
-    {
-        $this->beAnInstanceOf('AdminPanel\Symfony\AdminBundle\Tests\Doubles\MyCRUD');
-        $this->beConstructedWith([]);
-    }
-
-    public function it_is_initializable()
-    {
-        $this->shouldHaveType('AdminPanel\Symfony\AdminBundle\Admin\CRUD\AbstractCRUD');
-    }
-
-    public function it_is_admin_element()
-    {
-        $this->shouldHaveType('AdminPanel\Symfony\AdminBundle\Admin\Element');
-    }
-
-    public function it_have_default_route()
-    {
-        $this->getRoute()->shouldReturn('fsi_admin_list');
-    }
-
-    /**
-     * @param \FSi\Component\DataGrid\DataGridFactory $factory
-     * @throws \FSi\Component\DataGrid\Exception\DataGridColumnException
-     */
-    public function it_throw_exception_when_init_datagrid_does_not_return_instance_of_datagrid($factory)
-    {
-        $this->setDataGridFactory($factory);
-        $factory->createDataGrid(Argument::cetera())->willReturn(null);
-
-        $this->shouldThrow(new RuntimeException("initDataGrid should return instanceof FSi\\Component\\DataGrid\\DataGridInterface"))
-            ->during('createDataGrid');
-    }
-
-    /**
-     * @param \FSi\Component\DataGrid\DataGridFactory $factory
-     * @param \FSi\Component\DataGrid\DataGrid $datagrid
-     * @throws \FSi\Component\DataGrid\Exception\DataGridColumnException
-     * @throws \FSi\Component\DataGrid\Exception\UnexpectedTypeException
-     */
-    public function it_add_batch_column_to_datagrid_when_element_allow_delete_objects($factory, $datagrid)
-    {
-        $factory->createDataGrid('my_datagrid')->shouldBeCalled()->willReturn($datagrid);
-        $datagrid->hasColumnType('batch')->shouldBeCalled()->willReturn(false);
-        $datagrid->addColumn('batch', 'batch', [
-            'actions' => [
-                'delete' => [
-                    'route_name' => 'fsi_admin_batch',
-                    'additional_parameters' => ['element' => $this->getId()],
-                    'label' => 'crud.list.batch.delete'
-                ]
-            ],
-            'display_order' => -1000
-        ])->shouldBeCalled();
-
-        $this->setDataGridFactory($factory);
-
-        $this->createDataGrid()->shouldReturn($datagrid);
-    }
-
-    /**
-     * @param \FSi\Component\DataSource\DataSourceFactory $factory
-     * @throws \FSi\Component\DataSource\Exception\DataSourceException
-     */
-    public function it_throw_exception_when_init_datasource_does_not_return_instance_of_datasource($factory)
-    {
-        $this->setDataSourceFactory($factory);
-        $factory->createDataSource(Argument::cetera())->willReturn(null);
-
-        $this->shouldThrow(new RuntimeException("initDataSource should return instanceof FSi\\Component\\DataSource\\DataSourceInterface"))
-            ->during('createDataSource');
-    }
-
-    public function it_has_default_options_values()
-    {
-        $options = $this->getOptions();
-        $options->shouldHaveKey('allow_delete');
-        $options->shouldHaveKey('template_crud_list');
-        $options->shouldHaveKey('template_list');
-        $options['allow_delete']->shouldBe(true);
-        $options['template_crud_list']->shouldBe(null);
-        $options['template_list']->shouldBe(null);
-    }
-}
