--- src/AdminPanel/Symfony/AdminBundle/Admin/CRUD/CRUDElement.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AdminPanel\Symfony\AdminBundle\Admin\CRUD;
-
-interface CRUDElement extends ListElement, DeleteElement
-{
-}
