--- src/AdminPanel/Symfony/AdminBundle/Admin/CRUD/CRUDInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AdminPanel\Symfony\AdminBundle\Admin\CRUD;
-
-/**
- * @deprecated Deprecated since version 1.1, to be removed in 1.2. Use
- *             AdminPanel\Symfony\AdminBundle\Admin\CRUD\CRUDElement instead.
- */
-interface CRUDInterface extends CRUDElement
-{
-}
