--- src/AdminPanel/Symfony/AdminBundle/Doctrine/Admin/CRUDElement.php
@@ -1,29 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AdminPanel\Symfony\AdminBundle\Doctrine\Admin;
-
-use AdminPanel\Symfony\AdminBundle\Admin\CRUD\AbstractCRUD;
-
-abstract class CRUDElement extends AbstractCRUD implements Element
-{
-    use DataIndexerElementImpl;
-
-    /**
-     * {@inheritdoc}
-     */
-    public function saveDataGrid()
-    {
-        $this->getObjectManager()->flush();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function delete($object)
-    {
-        $this->getObjectManager()->remove($object);
-        $this->getObjectManager()->flush();
-    }
-}
