+++ src/AdminPanel/Symfony/AdminBundle/Resources/config/context/list.xml
@@ -6,31 +6,18 @@
 
     <parameters>
         <parameter key="admin.context.list.class">AdminPanel\Symfony\AdminBundle\Admin\CRUD\Context\ListElementContext</parameter>
-        <parameter key="admin.context.crud_list.class">AdminPanel\Symfony\AdminBundle\Admin\CRUD\Context\CRUDListElementContext</parameter>
         <parameter key="admin.context.list.request_handler.datasource_bind_parameters.class">AdminPanel\Symfony\AdminBundle\Admin\CRUD\Context\Request\DataSourceBindParametersHandler</parameter>
         <parameter key="admin.context.list.request_handler.datagrid_set_data.class">AdminPanel\Symfony\AdminBundle\Admin\CRUD\Context\Request\DataGridSetDataHandler</parameter>
         <parameter key="admin.context.list.request_handler.datagrid_bind_data.class">AdminPanel\Symfony\AdminBundle\Admin\CRUD\Context\Request\DataGridBindDataHandler</parameter>
     </parameters>
 
     <services>
-        <service id="admin.context.crud_list" class="%admin.context.crud_list.class%">
-            <argument type="collection">
-                <argument type="service" id="admin.context.list.request_handler.datasource_bind_params"/>
-                <argument type="service" id="admin.context.list.request_handler.datagrid_set_data"/>
-                <argument type="service" id="admin.context.list.request_handler.datagrid_bind_data"/>
-            </argument>
-            <argument type="string">%admin.templates.crud_list%</argument>
-            <argument type="service" id="form.factory"/>
-            <tag name="admin.context"/>
-        </service>
-
         <service id="admin.context.list" class="%admin.context.list.class%">
             <argument type="collection">
                 <argument type="service" id="admin.context.list.request_handler.datasource_bind_params"/>
                 <argument type="service" id="admin.context.list.request_handler.datagrid_set_data"/>
                 <argument type="service" id="admin.context.list.request_handler.datagrid_bind_data"/>
             </argument>
-            <argument type="service" id="form.factory"/>
             <tag name="admin.context"/>
         </service>
