--- src/AdminPanel/Symfony/AdminBundle/Tests/Doubles/Doctrine/MyCrudElement.php
@@ -1,37 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace AdminPanel\Symfony\AdminBundle\Tests\Doubles\Doctrine;
-
-use AdminPanel\Symfony\AdminBundle\Doctrine\Admin\CRUDElement;
-use FSi\Component\DataGrid\DataGridFactoryInterface;
-use FSi\Component\DataSource\DataSourceFactoryInterface;
-use Symfony\Component\Form\FormFactoryInterface;
-
-class MyCrudElement extends CRUDElement
-{
-    /**
-     * {@inheritdoc}
-     */
-    public function getClassName()
-    {
-        return 'FSiDemoBundle:Entity';
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getId()
-    {
-        return 'my_entity';
-    }
-
-    protected function initDataGrid(DataGridFactoryInterface $factory)
-    {
-    }
-
-    protected function initDataSource(DataSourceFactoryInterface $factory)
-    {
-    }
-}
