+++ src/AdminPanel/Symfony/AdminBundle/Tests/Doubles/MyCRUD.php
@@ -4,12 +4,12 @@ declare(strict_types=1);
 
 namespace AdminPanel\Symfony\AdminBundle\Tests\Doubles;
 
-use AdminPanel\Symfony\AdminBundle\Admin\CRUD\AbstractCRUD;
+use AdminPanel\Symfony\AdminBundle\Admin\CRUD\GenericListElement;
 use FSi\Component\DataGrid\DataGridFactoryInterface;
 use FSi\Component\DataSource\DataSourceFactoryInterface;
 use Symfony\Component\Form\FormFactoryInterface;
 
-class MyCRUD extends AbstractCRUD
+class MyCRUD extends GenericListElement
 {
     /**
      * {@inheritdoc}
