+++ tests/AdminPanel/Symfony/AdminBundle/Tests/Functional/Element/CustomTemplateUserElement.php
@@ -4,13 +4,13 @@ declare (strict_types = 1);
 
 namespace AdminPanel\Symfony\AdminBundle\Tests\Functional\Element;
 
-use AdminPanel\Symfony\AdminBundle\Doctrine\Admin\CRUDElement;
+use AdminPanel\Symfony\AdminBundle\Doctrine\Admin\ListElement;
 use AdminPanel\Symfony\AdminBundle\Tests\Functional\Entity\User;
 use FSi\Component\DataGrid\DataGridFactoryInterface;
 use FSi\Component\DataSource\DataSourceFactoryInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-final class CustomTemplateUserElement extends CRUDElement
+final class CustomTemplateUserElement extends ListElement
 {
     /**
      * Initialize DataGrid.
